% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KewlNewPkg.R
\name{KewlNewPkg}
\alias{KewlNewPkg}
\title{KewNewPkb}
\usage{
KewlNewPkg(x, y)
}
\arguments{
\item{x}{a n p matrix of n observation}

\item{y}{is another vector}
}
\value{
a vercot
}
\description{
KewNewPkb
}
\details{
This is an example function named 'hello'
which prints 'Hello, world!'.
}
\author{
mutecyphetr
}
\seealso{
\code{lm}
}

