% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KewlNewPkg.R
\name{predict10}
\alias{predict10}
\title{Prediction with love
takes code from \code{KewlNewPkg}}
\usage{
predict10(X, b)
}
\arguments{
\item{X}{an x by 10 matrix in the window}

\item{b}{a small dog with a waggy tail that no one wants}
}
\value{
a numberic vector with love
}
\description{
Takes coefficients and has love for them
}

